chore: refactor extract receipt function
